+++ CHANGELOG.md
@@ -18,6 +18,7 @@ Attention: don't forget to add the flag for F-Droid before release
 - [Feature] Add error display into remote controls screens
 - [Feature] Add new icons for remote-controls
 - [Feature] Add experimental option to enable remote controls
+- [Feature] Better information during synchronization
 - [Refactor] Load RemoteControls from flipper, emulating animation
 - [Refactor] Update to Kotlin 2.0
 - [Refactor] Replace Ktorfit with Ktor requests in remote-controls